Negative balance warning in a Section 104 holding
Since "same day" and "bed and breakfast" matched transactions are excluded from a Section 104 holding, it's possible (and valid) that a Withdrawal transaction can temporarily make the Section 104 holdings balance go negative.

The negative balance warning in the audit function remains, as this would indicate a real issue with your data.