Changed the append option so it now appends the original data as extra columns in the CSV output
Previously it was the other way around, with the transaction records appended to the end of the original data.

The appended format can now also be used by the bittytax tool, the additional data is just ignored. For this to work stdin must be encoded as utf-8 in Python 2.7, and also any empty rows need to be ignored.