Added Excel support for both the converter and accounting tools
bitttytax_conv

The —format argument now includes EXCEL as an output format in addition to CSV and RECAP. Excel is now the default instead of CSV.

The Excel spreadsheet created will contain a different worksheet for each of the converted data files. Data files of the same format are aggregated together.

The numbers in the worksheet(s) are formatted to ensure all decimal places are displayed correctly, numbers which exceed Excel’s 15-digits of precision are stored as text.

The transaction type column contains data validation so that only the correct values can be entered (i.e. Deposit, Mining, Income, Gift-Received) or selected from the dropdown menu.

The following arguments are not applicable to the EXCEL format.

1. -a (--append) - all original data is always appended to provide traceability to the created transaction records
2. -nh (--noheader) - each worksheet always contains a header row
3. -s (--sort) - each worksheet is always sorted by timestamp

A new argument --duplicates has been added to remove any duplicate input rows across data files of the same format. This is useful if you've had to repeatedly downloaded your transaction history from an exchange causing some overlap. It can be used for the CSV output format as well.

-o <output_filename>

The default filename for the EXCEL output is BittyTax_Records.xlsx.

If you want to use a different filename you can use the -o argument to specify this.

The -o argument can also be used for CSV files, by default the CSV output is to the terminal window (i.e. standard out), but by specifying an output filename the CSV data can be written directly to a file.

The bittytax conversion tool will never overwrite a file. If the filename already exists it will add a sequence number to the end (i.e. BittyTax_Records-2.xlsx, BittyTax_Records-3.xlsx, etc).

bittytax

The bittytax accounting tool now supports loading transaction records directly from an Excel file in addition to the existing CSV format.

Validation of the input file has been improved, instead of exiting at the first error, all transaction records will be validated and a full error report produced.

Each transaction record is allocated a unique TID (Transaction ID), these are in time ordered sequence.

The debug output has been updated to reflect that transaction records can be imported from both Excel or CSV files. For Excel files the worksheet name is given in addition to the row number.