Improved fee handling, including 3rd asset fees
Instead of fees being combined with the buy or sell's quantity/value, fees are now accounted for separately. Each Buy and Sell transaction contains a separate fee value in addition to the cost/proceeds.

If a fee is paid using a crypoasset, this disposal is recorded by the addition of a Spend transaction. This allows for fees to be paid using any cryptoasset, removing the limitation that the fee has to match the buy or sell asset.

Note: Spend transactions added for the fees of any transfer transactions (Deposit and Withdrawal) are not taxable so do not count as a disposal.

In the tax summary report the fees are listed in a separate column and taken into consideration when calculating the gain (or loss).

Resolves: #1
See also: #2