Price Data Report
Added appendix to the tax report PDF which contains a summary of all the historic price data used by BittyTax to value your specific cryptoassets.

The price data is split into financial years and contains the price in GBP (as well as the price in BTC where applicable) for each cryptoasset on a specific date, because counter-asset pricing is used for the valuation of trades, not every cryptoasset you own will appear in this table.

The cryptoasset name derived from the symbol name is also shown (i.e. Bitcoin for BTC), this is important to check to make sure your cryptoassets have been identified and priced correctly.

The name of the price data source used, along with a link to the actual API request used to retrieve the price data is also provided for reference.