Current Holdings Report and other report updates
Added a current holdings report to the appendix section at the end of the PDF document to show valuations for each token held.

Changed price data report so it's also included when a report is generated for a specific tax year.

Negative gains (i.e. a loss) are highlighed in red in the capital gains and current holdings reports.

Reworked the audit methods and added a html template for future audit reporting.