Timestamps normalized to local time only for Buy/Sell transactions
For transaction records the timestamps are no longer converted to local time, it's only done for Buy/Sell transactions which will be used for tax calculations.

This is to make it visible in the debug log what the original timestamp was and what it is when it's been normalized to local time.