Make the terminal output pretty
The logger output has been replaced by a more user friendly and more pretty output.

The colorama package is used to provide colour output on both Windows and Mac/Linux.

The tqdm package is used to display progress bars while the accounting tool is progressing transactions. Due to an issue with tqdm (https://github.com/tqdm/tqdm/issues/777) which creates extra new lines if any additional output is made during the progress bar, the setup.py has the tqdm package set to 'tqdm>=4.11.2,<4.25.0' as a workaround.

A progress "spinner" has also been added while the PDF report is being generated as this can be slow for large amounts of data.

The logger is no longer used.