Asset tool: new tool to list/search which assets have price data available
This tool is to be used in combination with the price tool. It lets you search all of the available data sources to see if a specific asset is availble. This is useful as some assets share the same symbol name, so it's important you use the correct data source.

Once you have identified the correct data source, you can update your bittytax.conf file to make "data_source_select" point to the correct data source for that specific symbol name, and then use the price tool to test it.

You can also choose to not specify any asset to get a complete list of all available assets, approx. 10,000 if you exclude duplicates.

usage: bittytax_asset [-h] [-v] [-d]
                      [-ds {COINDESK,COINGECKO,COINPAPRIKA,CRYPTOCOMPARE,EXCHANGERATESAPI,RATESAPI}]
                      [--duplicates]
                      [asset]