Etherscan parser: added ERC-20 tokens and ERC-721 NFTs exports
The etherscan parser has been updated to support exports of ERC-20 tokens and ERC-721 NFTs.

For these parsers to work the filename of these export files must contain your Ethereum address (which it does by default), this is so that transactions can be identfied as being sent or received correctly.

A new exception (UnknownAddressError) will be raised if the Ethereum address cannot be idenified, causing the parser to exit.

The filename is now passed as a new parameter to each "row_handler" and "all_handler" parsers, which can use it if required, only used by the Etherscan parser currently.