Qt Wallet parser: get symbol name from "Amount" if available
Some Qt Wallet data exports contain the symbol name with amount (ie. 1000 WDC). A regrex is used to match this, and if a symbol name is found this is used instead of forcing the user to specific one using the -ca command line option.

The -ca option now takes precedence over any symbol name found in the data file.