Conversion tool: Improved table formatting for Excel output
Added a table name which matches the parser name as used for the worksheet. Allows you to easily select the correct worksheet via its table name.

Added the column names correctly for the table, this requires any clash in names between the transaction record columns and the raw data columns to be modified if necessary (i.e. Timestamp, would become Timestamp2).