Accounting tool: validate tax year
The -ty or --taxyear argumnet is now validated to make sure it is supported (i.e. that CGT data exists for that year).

Added CGT data for tax year 2009, in case you were a very early adopter!

A warning is produced if your transactions records contain an unsupported tax year, might happen if you are running an old version of the tool.