Transaction notes feature branch
An optional note can now be added per transaction, taken from the last column after the timestamp.

This note shoud describe what the transaction is for, for example, for a deposit or withdrawal it could indicate which wallet the transfer is going to/from, to describe what the spend was for, or what the income was from.

The transaction note will appear in the logging to help find any issues.

The note will also be used as a description field in the income tax report to make tracking income sources easier.