Price tool: Added -ds ALL option to search all data sources
If you specify ALL with the -ds data source argument it will search all data sources for the asset you have specified and return any matches found. This is
useful if you are trying to find a price which is missing in the price data appendix in the tax report.

Extra validation has also been added for any quantity argument specified (-q or --quantity).
Extra validation of data source json responses has been added.

The output of the price tool has been improved to show which asset prices have ben used to calculate the price along with their sources, previously this was only visable with debug enabled (-d).

For assets which only have a BTC price (not GBP) the BTC price is retrieved first, followed by the BTC/GBP price to allow conversion. Previously it was the other way around.