Price tool: updated with list command to replace asset tool
The price tool has a new command argument added, this allows the tool to provide different functions which can have unique ar
guments.

The commands are: latest, historic and list:

    bittytax_price latest asset [quantity]
    bittytax_price historic asset date [quantity]
    bittytax_price list [asset]

The latest and historic commands provide the same functions as previously, the only different is that the quantity amount is now optional instead of having to be specified with the -q or --quantity argument.

A new list command has been added to provide the same function as the Asset tool (now removed), although simplied. If an asset is specified it will search all data sources and return any matches, otherwise without an asset, it will return all supported assets listed in alphabetical order.

The price tool now retrieves the asset price (in BTC) first, before retrieving the GBP price of BTC, this will also be reflected in the debug trace for the Accounting tool.

The output of the price tool will show each of these separate price retrievals, and their appropriate data sources, previousl
y this was only visible with debug enabled.