Features/kraken parser (#32)
Conversion tool: added parser for Kraken

Added a new parser for Kraken exchange. This parser supports both the "trades" and "ledger" exports. The ledger data file is used only for deposits/withdrawals, trades are ignored and taken from the trades data file which contains more detail.

Kraken uses non-standard asset names, some fiat ones are prefixed with Z and some crypto with X. Also the trading pair in the trades report does not include a seperator so it's difficult to split it the two assets.

To help do this, a script 'kraken_assets.py' has been added that uses the Kraken API to retrieve the asset names and generate constants which the parser will use.

Fix for #28 (Kraken format is not recognised).