Accounting tool: export option for transaction records
The '--export' command option has been added, this will export your transaction records with all the price data which BittyTax is using.

The file is in CSV format, the default filename is BittyTax_Export.csv. The file is ordered in chronological order, the same order that the accounting tool will process them.

This export might be useful if you wanted to verify BittyTax tax calculations using a different tool.