Binance parser: added cash deposit/withdrawals and statements
The Binance parser has been updated to support cash deposit and withdrawal exports in addition to crypto.

Since the header format is the same for both deposit and withdrawal exports, the parser now uses the filename to determine this (must contain deposit or withdrawal, case-insensitive). Previously it was using a fee of zero to identity deposits from withdrawals, but this was not working 100% of the time, as some cryptoassets have zero withdrawal fees (i.e. NEO).

The parser also now supports the “Statement” export, which contains details of all your transactions. Only those transactions not included elsewhere, i.e. distributions, BNB conversions, etc, will be converted into transaction records, so there is no duplication.

Note, for BNB conversions of multiple cryptoassets at the same time, the statement does not breakdown the BNB quantity per asset (it only gives a total), for these the parser will leave the BNB quantity empty, it's then up to the user to manually enter is data from the website (https://www.binance.com/en/my/wallet/history/bnbconvert). For this reason, it is recommended that you convert to BNB one-at-a-time.