coinbase: fix the missing fiat currency deposits in transactions
coinbase doesn't show the fiat currency deposits in transactions when a crypto is bought using debit or credit card. (coinbase pro seems to be showing this well)

However the notes column in the transaction show the amount used to buy the crypto and this can be used to make the assumption that the deposit has taken place.