Conversion tool: microsecond timestamps in Excel output
The Excel datetime can only be formatted to millseconds (i.e. 'yyyy-mm-dd hh:mm:ss.000'), if a data file contains microsecond timestamps (i.e. Kraken Trades) the timestamp is stored as a string format instead so no data is lost.

If a timestamp does not contain milli/microseconds the datetime cell is formatted as 'yyyy-mm-dd hh:mm:ss'.

The timestamp format used is consistant for the whole of the worksheet.