Price/Accounting tool: support for duplicate symbol names
The price and accounting tools have been updated to support duplicate symbol names, this currently only applies to the CoinGecko and CoinPaprika data sources.

You can use the price tool list command to see all assets, including any with duplicate symbols, together with their unique asset IDs. The default asset is indicated with a yellow ‘<-‘ arrow.

If the default asset ID selected for that symbol is not the one you require, it can be changed in the bittytax.conf file, in a similar way to how you specify the data source to be used for an asset. You can specify the asset ID as well, using ‘:’ colon as a separator.

For example:

data_source_select: {
    ...
    ‘FARM’ : [‘CoinGecko:harvest-finance'],
    ...
}

If you wanted to use both of the assets with the same symbol you could add it in the same way but use a different symbol.

For example:

data_source_select: {
    ...
    ‘FARMB’ : [‘CoinGecko:seed2need'],
    ...
}

The price tool's list command now has a search option (-s). This will do a simple case-insensitive search of all symbol and asset names in the list.

The list command also has the -ds option added, to select which data source to list the assets for.

When the price tool is displaying prices for a specific data source (-ds) or ‘ALL' data sources, duplicate symbols are now included, and the default indicated with the yellow ‘<-‘ arrow.