Accounting tool: added tax rules for UK companies
A new command option has been added to the accounting tool (—taxrules) to let you choose which tax rules bittytax should use when calculating taxes.

You can choose either UK_INDIVIDUAL (default) or UK_COMPANY_XXX, where XXX is the month which starts the company's financial year, i.e. JAN, FEB, etc.

For individuals the financial year starts 6th April, for a company the financial year will start on the 1st of the month specified.

A difference between the rules for individuals and companies is that the 30 day “bed and breakfast” rule is replaced with a 10 day rule instead. See https://www.gov.uk/government/publications/tax-on-cryptoassets/cryptoassets-tax-for-businesses#pooling-rules-exceptions-for-companies.

The tax report for UK companies will show a corporation tax estimate instead of a capital gains tax one. Estimates are included for both the “small profits rate” (when applicable) and the “main rate”, the percentage rate(s) used in the calculation are also shown. Two different rates can be used if the company's financial year straddles a 1st April rate change.

The tax report now shows the date range for the financial year, i.e. "Tax Year - 2020/21 (Apr 6th 2020 to Apr 5th 2021)”, also the percentage rate used for the capital gains calculation is shown.