Added local currency support
As we are getting more international users, I've decided to add local currency support as a hidden easter egg.

BittyTax currently only supports tax rules for the UK, but the price tool and conv tool are useful for all countries/currencies.

I've removed all hardcoded references to GBP, and there is now a new config parameter "local_currency", which can be set, it defaults to GBP if not defined.

This commit fixes #89 (Print report in EUR).