New config "transfer_fee_allowable_cost” added
This commit builds on the changes in commit ec7feaeeb08e7e9616c434357901788154118a66 which added the "transfer_fee_disposal" config parameter.

If "transfer_fee_disposal" is True, the new "transfer_fee_allowable_cost" parameter determines if the disposal is also an allowance cost for transfer fees (i.e. deposits/withdrawals).

The reason for adding this parameter is because the latest guidance from HMRC isn't clear, so this gives the user the option to choose, the default is False, so transfer fees are not an allowable cost. See #56 for discussion on this issue.

This new config parameter applies to ALL deposits/withdrawals, these could be between your own hot/cold wallets, or to/from your hosted wallet on an exchange, BittyTax makes no differentiation.