Conversion tool: better error handling for IOError
The conversion tool will now warn you if the file is missing, or if it's a directory. Useful if you are using wildcards in filenames.

This also fixes the missing "File could not be read" warning for missing files, introduced by commit c43dbbf826e05685e1c6635c70ad852a36f93774.