Accounting tool: don't store fixed value for transfers
For deposits/withdrawals, the Buy/Sell Value is not required, this is because transfers (between your own wallets) are not taxable events.

If a fixed value is specified, this will now NOT be store by bittytax, so will not be displayed in the debug logging. Hopefully this will make it clearer to the user that the fixed value is not required.