Accounting tool: The "ten day" rule for companies, should match the buy to sell, not sell to buy-back.
The "ten day" rule for companies has been corrected. Previously it was following the same logic as the "bed and breakfast" rule, so matching sells to buy-backs, when in fact it should be matching buys to sells, as stated in teh HMNRC guidance.

This change has been successfully tested using the HMRC examples here https://www.gov.uk/hmrc-internal-manuals/cryptoassets-manual/crypto41350/.

As part of this change, the "same day" rule logic has also been changed to match buy to sell, instead of sell to buy-back. It makes no difference to the calculations in this case, but is more intuitive in the log output.

If you have been using BittyTax with the --taxrules UK_COMPANY_XXX option, please re-generate your tax reports to see if this fix has impacted your tax calculations.

A side affect of this change is that we now have a generic FIFO method wnich can be used if a rule is not specified.

This change fixes #131.