Kraken parser: Trading pair split broken for XTZ/GBP
This is a fix for #124, the split_trading_pair function has been improved to prevent it incorrectly splitting the base and quote assets.

To prevent the XT/ZGBP split issue, it now checks that the base asset is 3 or more characters, the only acception being 'SC'.

The kraken_assets.py script has been updated to verify the split algorithm against all the current Kraken trading pairs.

This change also updates the QUOTE_ASSETS to align with the latest from Kraken.