Accounting tool: new transaction types Lost and Airdrop added
Added new transaction type 'Airdrop', this is purely descriptive, it behaves the same as a 'Gift-Received' so assumes that no income tax is payable for this specific airdrop. For airdrops which require income tax, please use 'Income' instead.

Another new transaction type 'Lost' has also been added. This is to be used if tokens have been lost or stolen (i.e. private keys are unrecoverable) and a "negligible value claim" has been reported/accepted by HMRC.

The lost tokens are treated as a disposal (for the quantity and value as agreed with HMRC) followed by a re-acquisition for the quantity lost, since technically you still own the tokens. If a value is not specified, a default disposal value of zero is assumed.

A new config item 'lost_buyback' has been added which is a boolean, this controls whether a 'Lost' transaction results in a re-acquision (default True) so could be used in the future for tax rules for other countries.

The Accointing and CoinTracking parsers have been updated to map to these new transaction types.

The Binance, Circle and Poloniex parsers have been updated to use Airdrop instead of Gift-Received where applicable.

The RECAP CSV format has been updated to map the new transaction types to the Recap.io equivalents.