Fix #155 (Error data file format unrecognised for Nexo)
The Nexo parser has been updated to support the new export format. This format includes a new field 'USD Equivalent', which is now used to populate the Buy/Sell values.

The 'Dividend' and 'Exchange' types are now mapped to DIVIDEND and TRADE respectively.

This parser still does not support the crypto loan types, including 'Interest' that is payable, identified as being negative. The crypto loan types are dependant on issue #135.