Binance parser: removed "Missing component data" error
For a "Small assets exchange BNB" operation it is possible that the convertion of token(s) to BNB can take multiple seconds to complete. For this reason, the "Missing component data" error has been removed if a BNB transaction cannot be found with the exact same timestamp.

Instead the parser will leave the "Buy Quantity" empty and the user will have to manually add these from the data on the website, https://www.binance.com/en/my/wallet/history/bnbconvert.

See #164 for more information.