Crypto.com parser: improved 'Native Amount' handling
The 'Native Amount' is now used even if the 'Native Currency' is not GBP (or the local currency). It will be converted into the local currency using DataParser.convert_currency().

Another change is that if the currency of the asset is already in GBP, a GBP value will not be populated as it serves no purpose.

Fix #170. The transaction type "council_node_deposit_created" has also been added, this is ignored as it is internal and not a taxable event.