Binance parser: handle symbols with leading digit
The Binance parser for the trades statement has been fixed so that it will handle base assets which have a leading digit (i.e. 1INCH).

The problem was that the regex was spliting the amount from the symbol, assuming the symbol would be only alpha characters. The soluton is for the parser to hold a list of those base assets which are prefixed with a digit and search for those first before attempting the split between digits and alpha characters.