Bitfinex parser: added ledger support
Added support for the ledger export format, this is used only for those transactions not covered by the Trades or Deposit/Withdrawals exports, i.e. referrals.

The trades parser has also been updated to round fees to 8 decimal places, and a fix for #197 to handle '-' in FEE PERC column, so the fee is ignored.