Etherscan merger and parser: fixes and changes
The Etherscan parser (and all its variants, i.e. BscScan, etc) have been fixed to filter any internal transactions with have failed (i.e. Status is not '0').

The parsers have also been updated to include the first 10 digits of the wallet address as part of the wallet name, i.e. "Ethereum-0x1a2b3c4d".

This unique wallet name is now used by the Etherscan merger to make sure it only merges transactions within the same wallet, previously you had to merge a single wallet at a time.

The Etherscan merger has been refactored so that all transactions are merged, not just those which appear in the transactions export but also those which only appear in the tokens/nfts export, this fixes #233.

The merger has also been updated so that transactions of the same asset are totaled up and reduced to a single deposit or withdrawal transaction type, at the moment this is only done for ETH transactions as doing it for all tokens breaks the logic which idenfies Staking income. This fixes #235.