Accounting tool: footer/page no. missing from PDF report
There is a bug in the xhtml2pdf package which prevents the footer with page no. appearing in the report.

This bug was introduced in v0.2.7 and is still not working in v0.2.8. As a workaround the setup.py has been updated to only install the previous working version v0.2.6.