Kraken parser: use Ledgers export for trades
The Kraken parser has been updated so that the Ledgers export (previously only used for Deposits/Withdrawals) is used for all transactions including trades, the Trades export is no longer required and is deprecated.

The Ledgers export is more accurate for trades, since some can have multiple fees which are not included in the Trades export, causing the data not to balance. It’s also possible to pay fees at a discount using Kraken's KFEE token, this is also only visible in the Ledgers export, see #91.

As well as handling trades, the Ledgers parser also supports these transaction types as defined here: https://support.kraken.com/hc/en-us/articles/360001169383-How-to-interpret-Ledger-history-fields

1. “staking”, see #153, #259
2. “invite bonus”, handled as a GIFT_RECEIVED
3. “transfer”, see #160, #87. Rebase is not currently supported
4. “spend”, handled the same as “trade”
5. ”receive”, handled the same as “trade”

The handling of Deposits/Withdrawals has been updated to support cancelled/failed transactions, see #149 and #287 respectively.

The types “margin trade”, “rollover”, “settled”, all relating to margin trading are not currently supported.

The type “adjustment” is not supported as I had no example data to go by.

The DataParser class has been updated to include a new attribute “deprecated” used to identify a parser which should no longer be used and point to the correct one.