Accointing parser: added new export format
The Accointing parser has been updated to support the new export format.

Any Accointing "classification" which cannot be mapped directly to a BittyTax transaction type is prefixed with an underscore, i.e. '_margin_gain'.

Any transaction types which could not be mapped are highlighed in red in the Excel spreadsheet, it's then up to the user to select the relevant bittytax transaction type from the dropdown validation menu.

This change fixes issue #219.