Koinly parser: allow unsupported types
Updated Koinly parser to allow unsupported types (i.e. those which cannot be mapped directly to a BittyTax transaction type) to be prefixed with an underscore, i.e. '_realized_gain'.

Any transaction types which could not be mapped are highlighed in red in the Excel spreadsheet, it's then up to the user to select the relevant bittytax transaction type from the dropdown validation menu.