CEX.IO parser: added wallet_buy/wallet_sell types
The CEX.IO parser has been upgraded to be an "all_handler" parser, this is required for wallet_buy/wallet_sell so the corresponding transaction can be found with a matching timestamp.

If more than one wallet_buy/wallet_sell combination for the same timestamp is found, an error will be produced.