Coinbase parser: check for Advanced Trades that trading pair matches the currency
The Coinbase parser has been updated, for Advanced Trades the "Notes" column contains the "on" trading pair used. If the quote part of the trading pair does no match the "for" fiat currency part of the "Note", a warning is produced and the relevant quantity is left blank, the user will have to populate this manually. This is because Coinbase does not provide the amount for the actual asset traded.

The Binance parser has also been updated to produce a simular warning if the BNB amount cannot be determined from the statements export, and will have to be added manually.