Conversion tool: removed 15-digit precision warning for CSV output
The 15-digit precision warning has been removed for CVS output, this was a hangover from earlier versions which didn't support Excel file output.

If you open a CSV file in Excel it will truncate high-precision numbers, whereas the Excel file output automically stores these numbers as strings to get around the issue.

Most people are now using the Excel file format so it's not really an issue, and the warning might confuse the user.

This change was requested by #308.