Config: replaced CoinGecko with CoinPaprika as secondary choice for crypto prices
Changing the default config (for new users) so that instead of using CoinGecko as the secondary choice, CoinPaprika is used for getting crypto prices.

There have been lots of reports of errors due to rate limiting with CoinGecko, see #291.

CoinPaprika seems to be better, it has a limit of 25k calls/month as opposed to 10k.