Accounting tool: added audit only (--audit) option
Added a new --audit option to the Accounting tool. This will just perform the audit function and nothing else. The report can be output to the terminal or PDF.

This is useful when you're still collating your transaction records, and trying to get everything to balance, it will prevent unnecessary price data API useage.

This fixes #97.