Coinbase parser: added new transaction types for transfers
Coinbase have updated their transaction history export to now include fiat deposits/withdrawals and also deposits/withdrawals which are relayed to/from Coinbase Pro.

For this reason the Coinbase/Coinbase Pro merge parser has been removed as it's no longer required.

Buy/Sells which are not in the local currency (i.e. EUR) are still shown as GBP so will not balance.

There appears to be a bug in the export, additional fiat "Sells" are added, between the traded fiat currency and the local fiat currency, even when they are the same, i.e. GBP to GBP, or GBP to EUR. These erroneous trades will be filtered by the accounting tool so cause no issue. Hopefully Coinbase will fix this.