Conversion tool: added parser for Gemini exchange
Added new parser for the Gemini exchange. The header rows in the exported spreadsheet is not fixed, as it contains running total columns for each asset you have purchased.

This required a change to DataParser to match header fields dynamically instead of fixed. For the header to match all fields must exist, and in the correct order, but don't need to be contiguous.