US tax rules corrections
Cashback and Fee-Rebate now included for Income tax.

The "trade_allowable_cost_type" config parameter set to default as 0 (buy asset).

For crypto-to-crypto trades the trade fee with be attributed to the buy asset, i.e. added to the cost basic. This will match how fees are calculated in the 1099-B documents provided by exchanges.