Accounting tool: handling of "VARIOUS" in TurboTax output formats
The TurboTax CSV/TXF formats don't allow you to specify if a disposal is a short or long term, instead they use the purchase date and date sold to calculate this.

This is fine, except if there are multiple purchase dates. Specifying the purchase date as "VARIOUS" may, or may not be recognised, either way you will need to manually categorise the diposal.

To aid the user, BittyTax will output "VARIOUS" in uppercase for long-term holdings, and lowercase "various" for short-term. This makes it easier to categorise the disposal correctly.