KuCoin parser: added "Spot Orders" and "Account History_Funding" files
Updated the KuCoin parser to support other files included in the bundle.

1. "Account History_Funding Account.csv" - this file is only used for Airdrops, all other transfers are duplicates.
2. "Spot Orders_Filled Orders.csv" - this file contains all your spot trades.

A new exception has been added DataFormatNotSupported, this is used by the KuCoin parser if it detects an "Account History_Trading Account.csv" file, which is not required/supported. This file has the same header format as the "Account History_Funding Account.csv" file, so the reason the parser has to detect this and error.