Accounting/Conversion tool: display output filenames with absolute paths
When BittyTax creates an output file it will be displayed with an absolute path. This is useful when the tools are run via drag and drop scripts, where it's not always clear in which directory the output file will be created.

The "is a directory" message has also been removed from the Conversion tool now that a directory can be specifed for the filename argument.