Kraken parser: handle "/" in trading pair
For the Trades export (which is deprecated, please use the Ledgers export), the trading pair now has a "/" seperator between the base and quote assets.

Added "spottofutures" which is an internal transfers which can be ignored.

Removed error if multiple transfers exist with the same refid as this can happen.