Accounting/Conversion tool: audit spreadsheet output to help with troubleshooting negative balances, and integrity check issues
If you run the accounting tool as audit only (--audit) a spreadsheet will be generated (BittyTax_Audit.xlsx) which shows for each transaction the wallet balance as it changes, as well as the total balance for that cryptoasset.

For each transaction you can also see the Transaction hash, Transaction source address, and Transaction destination address if these are provided in the raw data exported.

This is useful for pairing withdrawals with deposits to make sure there a no integrity check failures, you can use the Excel filtering to only show Withdrawal and Deposit transaction Type to make this process easier.

Clicking on the transaction Type will take you to the corresponding transaction record in your BittyTax_Records.xlsx file.