Accounting tool: fixed proceeds amount for a disposal which is both short and long term
For a disposal which has both short-term and long-term acquisitions, the proceeds should be the proportional amount for each entry, not the full amount.

The FIFO match warning (for a missing Buy), will now also be produced if only partial Buy matches are found (i.e. not enough to match the Sell amount entirely).