Accounting tool: new config parameter cost_basis_zero_if_missing
A new config parameter has been added "cost_basis_zero_if_missing" (default False) which allows BittyTax to add a Buy with a cost basis of zero for any Buy which cannot be matched to a Sell.

When this happens the integrity check will fail since more tokens have been added that originally accounted for during the audit. The difference should be the total amount of tokens added at zero cost.

A new option has been added to let you continue if the intergiry check fails without having to re-run the Accouting tool.