Accounting tool: fixed F8949 page duplication in PDF
If there were more than 2 pages of short/long-term disposals, subsequent pages were a duplicate of the previous page.

This was happening because PyPDF was using pointers to the same page, when adding new page, this is fixed by using reset_translation.

Also added a progress spinner while PDF is being generated.