Conversion tool: added parser and merge parser for CoinCorner
New parser added for CoinCorner, there are separate export data files for crypto and fiat, so requires a merge parser to re-construct the trades.

DataMerge required a small change to merge with only a single data parser, since CoinCorner's file format is the same for both.

CoinCorner do not include any fees in the exports files so will not balance exactly. It might be possible to hardcode or calculate the fees, please get in touch if you know what the fees should be.