Deribit parser: get uid from filename
Get the uid from the filename (if available) and use this to split positions and balances per uid.

This is a work in progress to implement the "position move" type correctly, ultimatley it will require a merge parser to process all data files (different uids) at once, not file by file.