Deribit parser: updates to handle partial closing of positions
Instead of checking if the position is zero for a "trade" or "settlement" Types, the parser now closes a position (partially or in full) if a "trade" Type's Side is set to "close sell" or "close buy".

For each open position, the total unrealised_pnl, funding_fees and trading_fees are accumulated. When a position is closed, the amount of realised_pnl/total fees is apportioned by the percentage of the position closed, with the remainder left to accumlate until the position is fully closed.

Other changes.

1. Timestamps are now calculated using a fraction of the UserSeq as microseconds. This ensures the transactions remain ordered correctly in the spreadsheet.
2. "settlement" Types for which there is no open position will raise an exception.
3. You can now specify the cryptoasset via the -ca argument if none can be identified from the filename.
4. "position move" will create an empty position if one does not already exist.

Observations:

1. A "close buy" or "close sell" can contain a "Cash Flow" amount which is not taken from the "Balance", since the parser includes this amount as an unrealised_pnl it means the parser will not balance correctly. Not sure how to resolve this yet.
2. For "position move" no balance changes are transfered across. For this reason the position that is moved is not apportioned.