Deribit parser: improvements to partial closing of positions
This fixes an issue with commit 9452de0102b9436232f112bcbb3170fd1374f5f9.

1. A "close buy" or "close sell" can contain a "Cash Flow" amount which is not taken from the "Balance", since the parser includes this amount as an unrealised_pnl it means the parser will not balance correctly. Not sure how to resolve this yet.

If "Cash Flow" is not taken from the "Balance" for a "close buy" or "close sell" trade, then it is not included in the accumulated unrealised_pnl.

A final settlement can happen after a position is closed, any "Cash Flow" here is treated as a seperate Margin Gain/Loss, this makes sure everything balances correctly.